crud de usuarios criado

+++ src/Controllers/UserController.js
@@ -14,10 +14,13 @@ module.exports = {
   async list(req, res) {
     try {
       const users = await User.find();
-      res.json({ users });
+      if (users === null) {
+        return res.status(401).json({ error: "Nenhum usuário cadastrado" });
+      }
+      return res.status(200).json({ users });
     } catch (error) {
       console.log(error);
-      res.json({ msg: "Problemas no servidor" });
+      return res.status(500).json({ msg: "Problemas no servidor" });
     }
   },
 
@@ -37,10 +40,13 @@ module.exports = {
       const user = await User.find({
         _id: userId
       });
-      res.json({ user });
+      if (user.lenght === 0) {
+        return res.status(401).json({ error: "Usuário não cadastrado" });
+      }
+      return res.status(200).json({ user });
     } catch (error) {
       console.log(error);
-      res.json({ msg: "Problemas no servidor" });
+      return res.status(500).json({ msg: "Problemas no servidor" });
     }
   },
 
@@ -57,12 +63,18 @@ module.exports = {
 
   async create(req, res) {
     try {
-      const { nome, sobrenome, ano } = req.body;
-      const user = await User.create({ nome, sobrenome, ano });
-      res.json({ user });
+      const { nome, idade, email, cargo } = req.body;
+      const userExist = await User.findOne({ email });
+      if (userExist !== null) {
+        return res
+          .status(401)
+          .json({ error: "Já existe um usuário com este e-mail" });
+      }
+      const user = await User.create({ nome, idade, email, cargo });
+      return res.status(201).json({ user });
     } catch (error) {
       console.log(error);
-      res.json({ msg: "Problemas no servidor" });
+      return res.status(500).json({ msg: "Problemas no servidor" });
     }
   },
 
@@ -82,24 +94,34 @@ module.exports = {
 
   async update(req, res) {
     try {
-      const { nome, sobrenome, ano } = req.body;
+      const { nome, idade, email, cargo } = req.body;
       const { userId } = req.params;
+
+      const userExist = await User.findById({ _id : userId });
+      
+      if (!userExist) {
+        return res
+          .status(401)
+          .json({ error: "Não existe este usuário" });
+      }
       const user = await User.findByIdAndUpdate(
         {
           _id: userId
         },
         {
           nome,
-          sobrenome
+          idade,
+          email,
+          cargo
         },
         {
           new: true
         }
       );
-      res.json({ user });
+      return res.status(200).json({ user });
     } catch (error) {
       console.log(error);
-      res.json({ msg: "Problemas no servidor" });
+      return res.status(500).json({ msg: "Problemas no servidor" });
     }
   },
 
@@ -116,13 +138,17 @@ module.exports = {
   async delete(req, res) {
     try {
       const { userId } = req.params;
+      const userExist = await User.findById({ _id: userId });
+      if (!userExist) {
+        return res.status(401).json({ error: "Usuário não cadastrado." });
+      }
       const user = await User.findByIdAndDelete({
         _id: userId
       });
-      res.json({ msg: "Usuário deletado com sucesso" });
+      return res.status(200).json({ msg: "Usuário deletado com sucesso" });
     } catch (error) {
       console.log(error);
-      res.json({ msg: "Problemas no servidor" });
+      return res.status(500).json({ msg: "Problemas no servidor" });
     }
   }
 };
